+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 
 Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image `hardisgroupcom/sfdx-hardis@beta`
 
+## [2.1.3] 2021-03-17
+
+- Fix hardis:retrieve:sources:dx when not in a DX project
+- Fix deloyment of Connected App in production
+- Display more options by page during interactive git add
+- Sort files to git add by group and manage preselection
+
 ## [2.1.2] 2021-03-14
 
 - Improve package installation